Create containers folder and move App to it
